refactor: Move game-id test logic into its own crate
